Polish added while making new Trailer
Animate Jerg Drone Spawn
Fixed some JSON formatting
AI will try to avoid getting on top of each other, even if it's a less optimal way of getting to the player. It leads to the player being surrounded which is actually even better!
Fix crash with potentially testing unicode string when it's actually an int
In debug, press f3 to hide UI (for screenshots)