Fixed issue with safe area on iOS
Changing the size of the "safe area" Control would sometimes allow buttons and label to overlap each other if the aspect ratio is too far from regular 16:9. This should be handled by the rescaling of the UI but changing the size of a Control for the safe area doesn't scale the UI. So instead I change the CanvasLayer transform to make the UI fit inside the safe area returned by Godot.